move paginatedSearch into new file
